Adding Test Cases
Added some preliminary tests — more to come. Changed the usage of
filterMinor slightly, introducing a flag to have it redirect rejected
graphs to stderr and changing the way flags are parsed to make that
more robust.  In findMinor, it will present the minor with letters if
the graph was input in pathlist format.